+++ xbmc/windows/GUIWindowFileManager.cpp
@@ -431,7 +431,7 @@ bool CGUIWindowFileManager::Update(int iList, const std::string &strDirectory)
   int iItem = GetSelectedItem(iList);
   std::string strSelectedItem = "";
 
-  if (iItem >= 0 && iItem < (int)m_vecItems[iList]->Size())
+  if (iItem >= 0 && iItem < m_vecItems[iList]->Size())
   {
     CFileItemPtr pItem = m_vecItems[iList]->Get(iItem);
     if (!pItem->IsParentFolder())
@@ -506,7 +506,7 @@ bool CGUIWindowFileManager::Update(int iList, const std::string &strDirectory)
   }
 
   // if we have a .tbn file, use itself as the thumb
-  for (int i = 0; i < (int)m_vecItems[iList]->Size(); i++)
+  for (int i = 0; i < m_vecItems[iList]->Size(); i++)
   {
     CFileItemPtr pItem = m_vecItems[iList]->Get(i);
     if (pItem->IsHD() &&
@@ -821,7 +821,7 @@ void CGUIWindowFileManager::Refresh()
   Update(0, m_Directory[0]->GetPath());
   Update(1, m_Directory[1]->GetPath());
 
-  while (nSel > (int)m_vecItems[iList]->Size()) {
+  while (nSel > m_vecItems[iList]->Size()) {
     nSel--;
 }
 
@@ -835,7 +835,7 @@ int CGUIWindowFileManager::GetSelectedItem(int iControl)
 }
   CGUIMessage msg(GUI_MSG_ITEM_SELECTED, GetID(), iControl + CONTROL_LEFT_LIST);
   if (OnMessage(msg)) {
-    return (int)msg.GetParam1();
+    return msg.GetParam1();
 }
   return -1;
 }
